chore(scripts): move manual verification scripts out of src
